Add support for test groups (#33)
Problem: LSpec is awesome, but when a codebase's test coverage is
good (as it should be 🙈), individual tests quickly become a wall
of text. Often in case of low-level tests that test polymorphic
functions, the descriptions are all basically the same except for
type names or some other minor details.

Solution: Haskell 'tasty' lib provides a test group concept, which is
basically a list of tests with a description and nice formatting.
Though LSpec is in many ways simpler than tasty, it would be nice to
steal this one construct.
Defined another constructor in the TestSeq inductive, consisting of a
description String field, a TestSeq of tests in the group, and the
optional 'next' sequence, same as before for individual tests.
Indentation (none for individual test sequences, two spaces for test
groups by default) was added to 'TestSeq.run' and 'TestSeq.runM'
functions to make test results print prettier.