fix: use correct bounds for `oneOf` and `elements`
The index of an array shouldn't go beyond its size-1.

This patch also removes the use of `Array.get!` from `Gen.lean` and uses `Array`
instead of `List` in `elements` because it does size computation and index-based
access in O(1) at runtime.

Further, it sets the `hi` value of the `DefaultRange` instance for `Nat` so that
it's within the `USize` range, for guaranteed memory optimization at runtime.